refactor: structure package
