+++ api-module/src/test/java/com/koa/apimodule/command/api/notice/service/NoticeFindUseCaseTest.java
@@ -2,6 +2,7 @@ package com.koa.apimodule.command.api.notice.service;
 
 import com.koa.commonmodule.testisolation.TestIsolation;
 import com.koa.coremodule.notice.application.dto.*;
+import com.koa.coremodule.notice.application.service.NoticeFindUseCase;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
