+++ api-module/src/test/java/com/koa/apimodule/command/api/report/service/ReportSaveUseCaseTest.java
@@ -2,6 +2,7 @@ package com.koa.apimodule.command.api.report.service;
 
 import com.koa.commonmodule.testisolation.TestIsolation;
 import com.koa.coremodule.report.application.dto.ReportRequest;
+import com.koa.coremodule.report.application.service.ReportSaveUseCase;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -9,15 +10,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 @SpringBootTest
 @TestIsolation
 @Sql(scripts = "/report_fixture.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
-class ReportUseCaseTest {
+class ReportSaveUseCaseTest {
 
     @Autowired
-    private ReportUseCase reportUseCase;
+    private ReportSaveUseCase reportSaveUseCase;
 
     @Nested
     @DisplayName("신고 작성")
@@ -33,7 +32,7 @@ class ReportUseCaseTest {
                     .content("신고 내용입니다.").build();
 
             //when
-            Long result = reportUseCase.createReport(reportRequest);
+            Long result = reportSaveUseCase.createReport(reportRequest);
 
             //then
             System.out.println(result);
