--- core-module/src/main/java/com/koa/coremodule/notice/application/exception/NoticeException.java
@@ -1,11 +1,0 @@
-package com.koa.coremodule.notice.application.exception;
-
-import com.koa.commonmodule.exception.BusinessException;
-import com.koa.commonmodule.exception.Error;
-
-public class NoticeException extends BusinessException {
-
-    public NoticeException(Error error) {
-        super(error);
-    }
-}
