+++ core-module/src/main/java/com/koa/coremodule/notice/application/service/NoticeFindUseCase.java
@@ -1,13 +1,13 @@
-package com.koa.apimodule.command.api.notice.service;
+package com.koa.coremodule.notice.application.service;
 
 import com.koa.coremodule.notice.application.dto.CurriculumListResponse;
 import com.koa.coremodule.notice.application.dto.CurriculumResponse;
 import com.koa.coremodule.notice.application.dto.NoticeListResponse;
-import com.koa.coremodule.notice.application.service.NoticeQueryService;
+import com.koa.coremodule.notice.domain.service.NoticeQueryService;
 import com.koa.coremodule.notice.domain.entity.Notice;
-import com.koa.coremodule.notice.mapper.NoticeMapper;
-import com.koa.coremodule.notice.repository.projection.CurriculumProjection;
-import com.koa.coremodule.notice.repository.projection.NoticeListProjection;
+import com.koa.coremodule.notice.application.mapper.NoticeMapper;
+import com.koa.coremodule.notice.domain.repository.projection.CurriculumProjection;
+import com.koa.coremodule.notice.domain.repository.projection.NoticeListProjection;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
