+++ core-module/src/main/java/com/koa/coremodule/notice/domain/repository/NoticeRepository.java
@@ -1,8 +1,7 @@
-package com.koa.coremodule.notice.repository;
+package com.koa.coremodule.notice.domain.repository;
 
 import com.koa.coremodule.notice.domain.entity.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.Repository;
 
 public interface NoticeRepository extends JpaRepository<Notice, Long>, NoticeDynamicRepository {
 }
