+++ core-module/src/main/java/com/koa/coremodule/notice/domain/repository/projection/CurriculumProjection.java
@@ -1,4 +1,4 @@
-package com.koa.coremodule.notice.repository.projection;
+package com.koa.coremodule.notice.domain.repository.projection;
 
 import com.querydsl.core.types.ConstructorExpression;
 import com.querydsl.core.types.Projections;
