+++ core-module/src/main/java/com/koa/coremodule/notice/domain/repository/projection/NoticeListProjection.java
@@ -1,33 +1,31 @@
-package com.koa.coremodule.notice.repository.projection;
+package com.koa.coremodule.notice.domain.repository.projection;
 
 import com.querydsl.core.types.ConstructorExpression;
 import com.querydsl.core.types.Projections;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import static com.koa.coremodule.notice.domain.entity.QCurriculum.curriculum;
 import static com.koa.coremodule.notice.domain.entity.QNotice.notice;
-import static com.koa.coremodule.notice.domain.entity.QNoticeImage.noticeImage;
 import static com.koa.coremodule.notice.domain.entity.QNoticeTeam.noticeTeam;
 
 @AllArgsConstructor
 @Getter
-public class NoticeDetailProjection {
+public class NoticeListProjection {
 
-    public static final ConstructorExpression<NoticeDetailProjection> CONSTRUCTOR_EXPRESSION =
-            Projections.constructor(NoticeDetailProjection.class,
+    public static final ConstructorExpression<NoticeListProjection> CONSTRUCTOR_EXPRESSION =
+            Projections.constructor(NoticeListProjection.class,
                     notice.id,
-                    notice.curriculum.curriculumName,
                     notice.title,
                     notice.content,
-                    noticeImage.imageUrl,
+                    curriculum.curriculumName,
                     noticeTeam.teamName
             );
 
     private Long noticeId;
-    private String curriculumName;
     private String title;
     private String content;
-    private String imageUrl;
+    private String curriculumName;
     private String teamName;
 
 }
