+++ core-module/src/main/java/com/koa/coremodule/report/domain/exception/ReportException.java
@@ -1,4 +1,4 @@
-package com.koa.coremodule.report.application.exception;
+package com.koa.coremodule.report.domain.exception;
 
 import com.koa.commonmodule.exception.BusinessException;
 import com.koa.commonmodule.exception.Error;
